requirements.txt
streamlit
PyPDF2
ollama
python-dotenv